Make State generic: move state outside of 'grpc' package
