+++ server/server.rs
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-use crate::service::grpc::state::ServerState;
+use crate::service::state::ServerState;
 use crate::service::{grpc, http};
 use crate::util::resolve_address;
 use crate::{
