+++ server/service/grpc/mod.rs
@@ -14,7 +14,6 @@ mod options;
 mod request_parser;
 mod response_builders;
 mod row;
-pub(crate) mod state;
 pub(crate) mod transaction_service;
 pub(crate) mod typedb_service;
 pub(crate) type ConnectionID = uuid::Bytes;
